# api/openapi.yaml
openapi: 3.1.0
info:
  title: betledger API
  version: 0.1.0
  description: |
    Wallet and wagering operations for game providers.

    Monetary amounts are always decimal strings with at most two decimal places,
    never JSON numbers, so that values never pass through floating point.

    Every response carries `X-Correlation-Id`. A value sent by the client is kept
    when it has 1 to 128 characters among letters, digits, `-`, `_` and `.`;
    otherwise a new one is generated.

    Business endpoints require an OAuth 2.0 access token issued by the identity
    provider through the `client_credentials` grant, sent as
    `Authorization: Bearer <token>`. The token must be signed by the provider,
    unexpired, and carry the `betledger-api` audience. Health endpoints are
    public.

    Authorization uses realm roles from the token: opening wallets requires
    `wallet-operator`, granted only to internal services, and sending operations
    requires `game-provider`. A provider acts only for the provider in its
    `provider_id` claim.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /wagering/transactions:
    post:
      tags: [Wagering]
      operationId: processWager
      summary: Send an operation from a game provider
      description: |
        Applies an operation to the wallet:

        | Kind | Movement | Reference |
        | --- | --- | --- |
        | `BET` | debit | none |
        | `WIN` | credit | optional, a `BET` of the same round |
        | `LOSS` | none; amount must be `0.00` | none |
        | `REFUND` | credit | required, a `BET` |
        | `ROLLBACK` | opposite of the reference | required, a `BET`, `WIN` or `REFUND` |

        The reference is found by `(providerId, referenceExternalTransactionId)`
        and must be `PROCESSED`, belong to the same player, wallet, currency and
        round, and, for a reversal, carry the same amount. An operation is
        reversed at most once: a `BET` accepts one `REFUND` or one `ROLLBACK`,
        never both. Disagreements are business refusals, answered with `422`.

        When the reference has not arrived yet, or is itself still pending, the
        operation is recorded as `PENDING_REFERENCE`, moves no money and is
        answered with `202`.

        Requires the `game-provider` role, and `providerId` must match the
        `provider_id` claim of the token; otherwise the answer is `403`.

        The operation, its ledger entry and its events are committed together,
        under a lock on the wallet, so concurrent operations on the same wallet
        are applied one at a time.

        **Idempotency.** Sending the same `Idempotency-Key` with an equivalent
        body returns the persisted result with `idempotentReplay: true`, including
        the balance observed originally, without applying anything again. The
        body is compared through a SHA-256 hash of its business fields in
        canonical JSON; the key and `X-Correlation-Id` are not part of it. Reusing
        the key with a different body, or resending the same
        `(providerId, externalTransactionId)` under another key, answers `409`.

        **Business refusals** — insufficient funds, a player who does not own the
        wallet, another currency — are persisted as `REJECTED` and answered with
        `422` and the result body, so that a retry returns the same outcome.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessWagerRequest'
            example:
              providerId: provider-a
              externalTransactionId: transaction-123
              playerId: 0192f28f-5dc0-7d58-bdb2-814ad6a0f4a1
              walletId: 0192f291-27dd-7d3f-8071-5f8685deef37
              roundId: round-987
              gameId: fortune-chimp
              kind: BET
              money:
                amount: '25.00'
                currency: BRL
      responses:
        '201':
          description: Operation applied now.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WagerResult'
              example:
                transactionId: 0192f298-345e-7e38-af88-e43f851a819d
                status: PROCESSED
                balance:
                  amount: '975.00'
                  currency: BRL
                idempotentReplay: false
        '202':
          description: |
            The operation was recorded as `PENDING_REFERENCE` and waits for its
            reference. It is retried in the background with exponential backoff
            and ends `PROCESSED` or `REJECTED` once the reference concludes, or
            `REJECTED` with `REFERENCE_NOT_FOUND` when the attempts run out.
            Replays answer the same way while it waits.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WagerResult'
              example:
                transactionId: 0192f298-345e-7e38-af88-e43f851a819d
                status: PENDING_REFERENCE
                idempotentReplay: false
        '200':
          description: Replay of an operation already applied; nothing was applied again.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WagerResult'
              example:
                transactionId: 0192f298-345e-7e38-af88-e43f851a819d
                status: PROCESSED
                balance:
                  amount: '975.00'
                  currency: BRL
                idempotentReplay: true
        '400':
          $ref: '#/components/responses/InvalidInput'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: The wallet does not exist.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: WALLET_NOT_FOUND
                  message: wallet 0192f291-27dd-7d3f-8071-5f8685deef37 not found
        '409':
          description: |
            The key was already used with a different body, or the operation was
            already sent under another key.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: IDEMPOTENCY_CONFLICT
                  message: idempotency key "provider-a:transaction-123" was already used with a different payload
        '422':
          description: |
            The operation was refused by a business rule and persisted as
            `REJECTED`. Replays of it answer the same way, with
            `idempotentReplay: true`.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WagerResult'
              example:
                transactionId: 0192f298-345e-7e38-af88-e43f851a819d
                status: REJECTED
                failureCode: INSUFFICIENT_FUNDS
                idempotentReplay: false
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /wagering/transactions/{transactionId}:
    get:
      tags: [Wagering]
      operationId: getWagerTransaction
      summary: Read an operation
      description: |
        Returns the operation, with its state, the balance observed when it was
        processed, the failure code of a refusal and how many attempts a pending
        reference already had. Requires the `game-provider` role; an operation of
        another provider answers `404`, so that a provider cannot learn that it
        exists.
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: The operation.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WagerTransaction'
        '400':
          $ref: '#/components/responses/InvalidInput'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/TransactionNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /providers/{providerId}/wagering/transactions/{externalTransactionId}:
    get:
      tags: [Wagering]
      operationId: getWagerTransactionByExternalId
      summary: Read an operation by the provider identifier
      description: |
        Returns the operation the provider identified with
        `externalTransactionId`. The `providerId` in the path must be the one in
        the token; another provider answers `403`.
      parameters:
        - name: providerId
          in: path
          required: true
          schema:
            type: string
        - name: externalTransactionId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: The operation.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WagerTransaction'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/TransactionNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token obtained from the identity provider with the
        `client_credentials` grant. Locally:

        ```sh
        curl -s http://localhost:8081/realms/betledger/protocol/openid-connect/token \
          -d grant_type=client_credentials \
          -d client_id=provider-a -d client_secret=provider-a-secret
        ```
  parameters:
...
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: Identifier that ties together logs and events of the request.
      schema:
        type: string
        pattern: '^[A-Za-z0-9._-]{1,128}$'
        example: req-123
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Identifies retries of the same operation within the provider. It may be
        built as `{providerId}:{externalTransactionId}`, but the server never
        replaces it with a computed one.
      schema:
        type: string
        minLength: 1
        example: provider-a:transaction-123
  headers:
    CorrelationId:
      description: The correlation identifier kept or generated for the request.
      schema:
        type: string
  responses:
    InvalidInput:
      description: The request is invalid and must be fixed before being sent again.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            invalidAmount:
              summary: Amount with more than two decimal places
              value:
                error:
                  code: INVALID_AMOUNT
                  message: monetary amount "10.123" exceeds the scale of 2 decimal places
            invalidInput:
              summary: Malformed body
              value:
                error:
                  code: INVALID_INPUT
                  message: 'malformed request body: unexpected EOF'
    Unauthenticated:
      description: The bearer token is missing, invalid, expired or issued for another audience.
      headers:
        WWW-Authenticate:
          schema:
            type: string
            example: Bearer realm="betledger"
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHENTICATED
              message: a bearer token is required
...
    TransactionNotFound:
      description: The operation does not exist, or belongs to another provider.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: TRANSACTION_NOT_FOUND
              message: operation "transaction-123" not found for provider provider-a
    Forbidden:
      description: |
        The caller is authenticated but lacks the role for the endpoint, or the
        operation names a provider other than the one in the token.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: FORBIDDEN
              message: the wallet-operator role is required
    ServiceUnavailable:
      description: A dependency is temporarily unavailable. The request can be retried.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: SERVICE_UNAVAILABLE
              message: service temporarily unavailable, retry later
    InternalError:
      description: Unexpected failure. Details are logged, never returned.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: INTERNAL_ERROR
              message: internal error
  schemas:
    MoneyInput:
      type: object
      description: Non-negative amount accepted in requests.
      required: [amount, currency]
      additionalProperties: false
      properties:
        amount:
          type: string
          description: Decimal string with at most two decimal places.
          pattern: '^[0-9]+(\.[0-9]{1,2})?$'
          example: '1000.00'
        currency:
          type: string
          description: ISO 4217 code.
          pattern: '^[A-Z]{3}$'
          example: BRL
    Money:
      type: object
      description: Amount returned by the service, always with two decimal places.
      required: [amount, currency]
      additionalProperties: false
      properties:
        amount:
          type: string
          pattern: '^-?[0-9]+\.[0-9]{2}$'
          example: '1000.00'
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          example: BRL
...
    ProcessWagerRequest:
      type: object
      required: [providerId, externalTransactionId, playerId, walletId, roundId, gameId, kind, money]
      additionalProperties: false
      properties:
        providerId:
          type: string
          minLength: 1
          description: Must match the `provider_id` claim of the token.
        externalTransactionId:
          type: string
          minLength: 1
          description: Unique per provider; the operation cannot be applied twice.
        playerId:
          type: string
          format: uuid
        walletId:
          type: string
          format: uuid
        roundId:
          type: string
          minLength: 1
        gameId:
          type: string
          minLength: 1
        kind:
          type: string
          enum: [BET, WIN, LOSS, REFUND, ROLLBACK]
        money:
          $ref: '#/components/schemas/MoneyInput'
        referenceExternalTransactionId:
          type: string
          description: |
            The provider's identifier of the referenced operation; required by
            `REFUND` and `ROLLBACK`, optional for `WIN`, refused for the others.
    WagerResult:
      type: object
      required: [transactionId, status, idempotentReplay]
      additionalProperties: false
      properties:
        transactionId:
          type: string
          format: uuid
        status:
          type: string
          enum: [PROCESSED, REJECTED, PENDING_REFERENCE]
        balance:
          $ref: '#/components/schemas/Money'
          description: Balance right after the operation was applied; present only when `PROCESSED`.
        failureCode:
          type: string
          description: |
            Why the operation was `REJECTED`; absent otherwise. One of
            `INSUFFICIENT_FUNDS`, `INSUFFICIENT_FUNDS_FOR_REVERSAL`,
            `WALLET_PLAYER_MISMATCH`, `CURRENCY_MISMATCH`, `REFERENCE_MISMATCH`,
            `REFERENCE_NOT_FOUND`, `REFERENCE_NOT_PROCESSED`, `REFERENCE_AMOUNT_MISMATCH` or
            `REFERENCE_ALREADY_REVERSED`.
          example: INSUFFICIENT_FUNDS
        idempotentReplay:
          type: boolean
          description: True when the result was already persisted and nothing was applied again.
...
    WagerTransaction:
      type: object
      required:
        [transactionId, providerId, externalTransactionId, kind, status, walletId, playerId, roundId, gameId,
         money, referenceAttempts]
      additionalProperties: false
      properties:
        transactionId:
          type: string
          format: uuid
        providerId:
          type: string
        externalTransactionId:
          type: string
        kind:
          type: string
          enum: [BET, WIN, LOSS, REFUND, ROLLBACK]
        status:
          type: string
          enum: [PENDING, PENDING_REFERENCE, PROCESSED, REJECTED, FAILED]
        walletId:
          type: string
          format: uuid
        playerId:
          type: string
          format: uuid
        roundId:
          type: string
        gameId:
          type: string
        money:
          $ref: '#/components/schemas/Money'
        balance:
          $ref: '#/components/schemas/Money'
          description: Balance observed when the operation was processed; absent otherwise.
        failureCode:
          type: string
          description: Why the operation was refused; absent otherwise.
        referenceExternalTransactionId:
          type: string
        referenceAttempts:
          type: integer
          minimum: 0
          description: Attempts that found no concluded reference.
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              description: |
                Stable failure code, part of the contract. Codes returned so far:
                `INVALID_INPUT`, `INVALID_AMOUNT`, `AMOUNT_OVERFLOW`,
                `TRANSACTION_KIND_NOT_ALLOWED`, `WALLET_ALREADY_EXISTS`,
                `WALLET_NOT_FOUND`, `IDEMPOTENCY_CONFLICT`, `UNAUTHENTICATED`, `FORBIDDEN`,
                `SERVICE_UNAVAILABLE`,
                `INTERNAL_ERROR`.
            message:
              type: string
              description: Human-readable description; not meant to be parsed.
...